requirements.txt
--index-url https://download.pytorch.org/whl/cpu
flask>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
pillow-heif>=0.15.0
